tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-name_two_values.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-name value with two list entries</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-name-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="IDENT, IDENT is a valid animation-name value">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes fade {
         from {
            opacity: 1;
         }
         to {
            opacity: 0;
         }
      }
      @keyframes slide {
         from {
            margin-left: 0;
         }
         to {
            margin-left: 100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
      }
      .test {
         animation-duration: 1s, 1s;
         animation-name: fade, slide;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var fadeEnds = 0;
         var slideEnds = 0;
         var passed = undefined;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         div.addEventListener('animationend', function(evt) {
            if (evt.animationName == 'fade') {
               fadeEnds++;
            } else if (evt.animationName == 'slide') {
               slideEnds++;
            } else {
               passed = false;
            }
            if (fadeEnds == 1 && slideEnds == 1) {
               clearTimeout(timeout);
               passed = (passed == undefined);
               test(function(){assert_true(passed);}, testName); done();
               log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
            }
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The block below should fade out while sliding right, then reappear.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
